+++ pywincffi/kernel32/file.py
